Edit: id 타입 int로 통일 및 컬럼명 통일

+++ src/main/java/com/example/pet/controller/ReviewController.java
@@ -60,7 +60,7 @@ public class ReviewController {
      */
 
     @GetMapping("/myPage/review/{memberId}")
-    public List<GetReviewDto> getReview(@PathVariable Long memberId){
+    public List<GetReviewDto> getReview(@PathVariable int memberId){
 
         return reviewService.getReviewList(memberId);
